Modified slightly Band_by_Band test
Added the option of choosing between comparing the power of detected peaks or the power of frequency band ranges. While I was writing examples, I realized that this would make it easier for users to choose between the two instead of modifying the class each time. By default it will give the power of frequency band ranges. If option=1, then it will compute the power of detected peaks. This might be more clear with an example which I am currently working on for GitHub. For now, there is an example in the notebook new_band_by_band_comparison.ipynb (in FOOOF_tests). Let me know if this modification suits you.